+++ clustering/cluster_graph_concepts.py
@@ -37,8 +37,10 @@ class ConceptGraphClusterer:
 
         # Adjacency list built from Graph
         db = self.build_adj_list()
-        with open(cache_path, "w", encoding="utf-8") as f:
-            json.dump(db, f, ensure_ascii=False, indent=4)
+
+        if os.path.exists(cache_path) and self.config['general']['should_cache']:
+            with open(cache_path, "w", encoding="utf-8") as f:
+                json.dump(db, f, ensure_ascii=False, indent=4)
 
         # Transitive closure
         self.transitive_closure(db)
